TumblrRestClient: miscellaneous parameter validation fixups
Add a _post_valid_options() method, which returns a list of valid
options for a given post type (which may be None, in which case the
common list of valid post options are returned). This lets us remove
the valid_options argument from _send_post() and fix up the parameter
validation in the reblog() and edit_post() methods.

While we're here, also add missing valid parameters to the edit_post
docstring.