Handle 302 responses same as other success codes
We previously only treated 200, 201 and 301 as successes, and everything
else as an error. Switch to treating all codes between 200 and 399 as
successes.

This was necessary due to a backend change where avatars are now
returned as 302 responses instead of 301, causing their entire json
payload to be returned instead of just the relevant portion.