input fields fixed
Input fields may have extra blank spaces which can cause the termination of the program due to wrong tokens/secret keys. Especially in the web browser in this link (https://www.tumblr.com/oauth/apps), when you double click on the secret key to select all of the key and copy, it selects extra three or more blank spaces which cause failure and termination of the program.